refactor: simplify global variable builder setup
